feat: finalize ADR-063 Phases 1 and 2 (dump/scan convergence, EntityId wiring)
Phase 1: migrate the PE/Mach-O `dump` CLI path onto the shared
execute_dump_request pipeline the same way the ELF path already was,
so both binary formats now execute through one typed pipeline.
handle_non_elf_dump is retired for the CLI's real dispatch but kept
defined for its own unit tests.

Phase 2: wire Change.entity_id at every remaining RecordType/EnumType/
Function/Variable-backed diff call site (diff_types.py,
diff_types_field_facts.py, diff_layout.py, diff_vtable_layout.py,
diff_symbols.py, diff_symbols_variables.py), following the established
old-preferred/new-fallback convention, with field/base-level findings
attributed to their containing record's identity. Two new leaf modules
(diff_types_bases_vtable.py, diff_symbols_anon_fields.py) split out of
diff_types.py/diff_symbols.py to stay under the architecture file-size
gates after the wiring's line growth.

Typedef/constant sites, DWARF/platform-only detectors, and the L5
source-graph's own identity scheme remain out of scope, each for a
documented, separate reason (no EntityId producer, no EntityId at that
tier yet, or an already-separate identity primitive) -- recorded
precisely in the ADR-063 status text and implementation plan rather
than silently left for a future reader to rediscover.
