+++ abicheck/diff_vtable_layout.py
@@ -172,6 +172,7 @@ def _diff_vtable_layout(old: AbiSnapshot, new: AbiSnapshot) -> list[Change]:
                     name=name,
                     old=", ".join(o.virtual_bases),
                     new=", ".join(n.virtual_bases),
+                    entity_id=o.entity_id or n.entity_id,
                 )
             )
 
@@ -201,6 +202,7 @@ def _diff_vtable_layout(old: AbiSnapshot, new: AbiSnapshot) -> list[Change]:
                         name=name,
                         old=", ".join(og) or "(none)",
                         new=", ".join(ng) or "(none)",
+                        entity_id=o.entity_id or n.entity_id,
                     )
                 )
