+++ architecture/modules.yaml
@@ -204,6 +204,7 @@
         "abicheck/diff_surface_metrics.py",
         "abicheck/diff_sycl.py",
         "abicheck/diff_symbols.py",
+        "abicheck/diff_symbols_anon_fields.py",
         "abicheck/diff_symbols_renames.py",
         "abicheck/diff_symbols_scalar.py",
         "abicheck/diff_symbols_variables.py",
@@ -212,6 +213,7 @@
         "abicheck/diff_type_spellings.py",
         "abicheck/diff_types.py",
         "abicheck/diff_types_abicc_parity.py",
+        "abicheck/diff_types_bases_vtable.py",
         "abicheck/diff_types_field_facts.py",
         "abicheck/diff_types_surface.py",
         "abicheck/diff_unnamed_types.py",
@@ -498,6 +500,7 @@
       "diff_surface_metrics.py",
       "diff_sycl.py",
       "diff_symbols.py",
+      "diff_symbols_anon_fields.py",
       "diff_symbols_renames.py",
       "diff_symbols_scalar.py",
       "diff_symbols_variables.py",
@@ -506,6 +509,7 @@
       "diff_type_spellings.py",
       "diff_types.py",
       "diff_types_abicc_parity.py",
+      "diff_types_bases_vtable.py",
       "diff_types_field_facts.py",
       "diff_types_surface.py",
       "diff_unnamed_types.py",
