+++ changelog.d/20260901_035444_noreply_adr_063_phases_1_2_t49ilm.md
@@ -1,0 +1,81 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+### Added
+
+- **`Change.entity_id` (ADR-063 Phase 2) now populated across the
+  layout/vtable-layout/variable detectors and the remaining `diff_symbols.py`
+  gaps** — `diff_layout.py`'s six layout-descriptor findings,
+  `diff_vtable_layout.py`'s two vtable-group findings,
+  `diff_symbols_variables.py`'s alignment/access-level findings, and
+  nine previously-unwired `diff_symbols.py` sites (`VAR_TYPE_CHANGED`,
+  the const-transition pair, `VAR_REMOVED`/`VAR_ADDED`,
+  `FIELD_ACCESS_CHANGED`, both `ANON_FIELD_CHANGED` branches, and
+  `VAR_DEPRECATED_ADDED`/`REMOVED`) all now carry the matched `RecordType`/
+  `Variable`'s resolved identity, following the existing old-preferred/
+  new-fallback convention. The three `CONSTANT_*` findings stay unwired —
+  `AbiSnapshot.constants` is a plain `dict[str, str]` with no parsed
+  declaration behind it to carry an `EntityId`. Still `compare=False`, so
+  no equality-based behavior changes.
+
+<!--
+### Changed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Deprecated
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Removed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Fixed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Performance
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Security
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Documentation
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
