+++ changelog.d/20260901_035804_noreply_adr_063_phases_1_2_t49ilm.md
@@ -1,0 +1,22 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+### Changed
+
+- **`Change.entity_id` now populated across the type-diff detectors** —
+  every struct/union/enum/typedef-family finding from `diff_types.py` and
+  `diff_types_field_facts.py` (type/field/base/vtable/enum-member changes)
+  now carries the matched `RecordType`/`EnumType`/`Function`'s `entity_id`
+  where one is available, continuing ADR-063 Phase 2's exhaustive
+  population beyond the function-diff path. Field-level findings attribute
+  to their containing record's identity (fields have no `EntityId` of
+  their own yet). `entity_id` remains `compare=False` and additive-only
+  (folded into `resolve_change_identity` as a non-primary alias), so no
+  existing suppression rule, canonical finding ID, or report shape
+  changes. Typedef/constant-family findings are unaffected — no
+  `EntityId` producer exists for them yet.
