+++ changelog.d/20260901_040135_noreply_adr_063_phases_1_2_t49ilm.md
@@ -1,0 +1,76 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+### Changed
+
+- **`abicheck dump`'s real PE/Mach-O run now executes through the same
+  typed `execute_dump_request` pipeline as ELF** (ADR-063 Phase 1,
+  completing the PR C migration) — `handle_non_elf_dump` is retired for
+  the `dump` CLI's real (non-`--dry-run`) dispatch, mirroring the earlier
+  ELF migration onto `frontends/cli/dump_execute.execute_dump_cli_run`; it
+  stays defined, unchanged, for its own direct unit tests. No CLI-visible
+  behavior change for a valid invocation. Verified via mock-based CLI/unit
+  tests only — no PE/Mach-O toolchain was available to verify a real
+  end-to-end run the way the ELF migration's own test corpus does.
+
+<!--
+### Added
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Deprecated
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Removed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Fixed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Performance
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Security
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Documentation
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
