+++ changelog.d/20260901_053124_noreply_adr_063_phases_1_2_t49ilm.md
@@ -1,0 +1,16 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+### Changed
+
+- **Split `diff_types.py`'s base-class and vtable diffing into
+  `diff_types_bases_vtable.py`** — a mechanical, behavior-preserving
+  extraction keeping `diff_types.py` under the AI-readiness file-size hard
+  cap after ADR-063 Phase 2's `entity_id` wiring; no detection logic,
+  `ChangeKind`, or CLI/report behavior changed.
+
