+++ changelog.d/20260901_060625_noreply_adr_063_phases_1_2_t49ilm.md
@@ -1,0 +1,78 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+<!--
+### Added
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+### Changed
+
+- **Split `diff_symbols.py`'s anonymous struct/union field diffing into
+  `diff_symbols_anon_fields.py`** — a mechanical, behavior-preserving
+  extraction keeping `diff_symbols.py` within its architecture-debt
+  no-growth baseline after ADR-063 Phase 2's `entity_id` wiring; no
+  detection logic, `ChangeKind`, or CLI/report behavior changed. Also
+  corrects `docs/contribute/adr/063-one-semantic-pipeline.md`'s Phase 1/2
+  status bullets to reflect that the PE/Mach-O `dump` migration and the
+  bulk of Phase 2's exhaustive `entity_id` population have landed, and
+  names the three site classes (typedefs/constants, DWARF/platform-only
+  detectors, the L5 source-graph's own identity scheme) that remain
+  deliberately unwired.
+
+<!--
+### Deprecated
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Removed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Fixed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Performance
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Security
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Documentation
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
