+++ scripts/fact_field_readers.py
@@ -209,7 +209,7 @@ KNOWN_UNMIGRATED_READERS: frozenset[str] = frozenset(
         "abicheck/diff_cxx_rules.py::_transitive_bases::virtual_bases::stack.extend((*rec.bases, *rec.virtual_bases))::rec.virtual_bases::1",
         "abicheck/diff_cxx_rules.py::virtual_method_addition::vtable::t_old.vtable != t_new.vtable::t_new.vtable::1",
         "abicheck/diff_cxx_rules.py::virtual_method_addition::vtable::t_old.vtable != t_new.vtable::t_old.vtable::1",
-        'abicheck/diff_layout.py::_check_vptr_introduced::vptr_offset_bits::[\n            make_change(\n                ChangeKind.VPTR_INTRODUCED,\n                symbol=name,\n                name=name,\n                old_value="non-polymorphic",\n                new_value=f"vptr@{new_rec.vptr_offset_bits}",\n            )\n        ]::new_rec.vptr_offset_bits::1',
+        'abicheck/diff_layout.py::_check_vptr_introduced::vptr_offset_bits::[\n            make_change(\n                ChangeKind.VPTR_INTRODUCED,\n                symbol=name,\n                name=name,\n                old_value="non-polymorphic",\n                new_value=f"vptr@{new_rec.vptr_offset_bits}",\n                entity_id=old_rec.entity_id or new_rec.entity_id,\n            )\n        ]::new_rec.vptr_offset_bits::1',
         "abicheck/diff_layout.py::_check_vptr_introduced::vptr_offset_bits::not old_rec.vtable\n        and old_rec.vptr_offset_bits is None\n        and new_rec.vtable\n        and new_rec.vptr_offset_bits is not None::new_rec.vptr_offset_bits::1",
         "abicheck/diff_layout.py::_check_vptr_introduced::vptr_offset_bits::not old_rec.vtable\n        and old_rec.vptr_offset_bits is None\n        and new_rec.vtable\n        and new_rec.vptr_offset_bits is not None::old_rec.vptr_offset_bits::1",
         "abicheck/diff_layout.py::_check_vptr_introduced::vtable::not old_rec.vtable\n        and old_rec.vptr_offset_bits is None\n        and new_rec.vtable\n        and new_rec.vptr_offset_bits is not None::new_rec.vtable::1",
@@ -223,46 +223,46 @@ KNOWN_UNMIGRATED_READERS: frozenset[str] = frozenset(
         "abicheck/diff_stdlib_impl.py::_public_by_value_type_closure::virtual_bases::(*record.bases, *record.virtual_bases)::record.virtual_bases::1",
         "abicheck/diff_time64.py::_fold_record_tokens::bases::list(rec.bases) + list(rec.virtual_bases)::rec.bases::1",
         "abicheck/diff_time64.py::_fold_record_tokens::virtual_bases::list(rec.bases) + list(rec.virtual_bases)::rec.virtual_bases::1",
-        "abicheck/diff_types.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_POSITION_CHANGED,\n                symbol=name,\n                name=name,\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n            )\n        )::t_new.bases::1",
-        "abicheck/diff_types.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_POSITION_CHANGED,\n                symbol=name,\n                name=name,\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n            )\n        )::t_old.bases::1",
-        'abicheck/diff_types.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.TYPE_BASE_CHANGED,\n                symbol=name,\n                description=f"Base classes changed: {name}",\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n            )\n        )::t_new.bases::1',
-        'abicheck/diff_types.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.TYPE_BASE_CHANGED,\n                symbol=name,\n                description=f"Base classes changed: {name}",\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n            )\n        )::t_old.bases::1',
-        "abicheck/diff_types.py::_diff_type_bases::bases::old_bases_set == new_bases_set and t_old.bases != t_new.bases::t_new.bases::1",
-        "abicheck/diff_types.py::_diff_type_bases::bases::old_bases_set == new_bases_set and t_old.bases != t_new.bases::t_old.bases::1",
-        "abicheck/diff_types.py::_diff_type_bases::bases::set(t_new.bases)::t_new.bases::1",
-        "abicheck/diff_types.py::_diff_type_bases::bases::set(t_old.bases)::t_old.bases::1",
-        'abicheck/diff_types.py::_diff_type_bases::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.TYPE_BASE_CHANGED,\n                    symbol=name,\n                    description=f"Virtual base classes changed: {name}",\n                    old_value=str(t_old.virtual_bases),\n                    new_value=str(t_new.virtual_bases),\n                )\n            )::t_new.virtual_bases::1',
-        'abicheck/diff_types.py::_diff_type_bases::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.TYPE_BASE_CHANGED,\n                    symbol=name,\n                    description=f"Virtual base classes changed: {name}",\n                    old_value=str(t_old.virtual_bases),\n                    new_value=str(t_new.virtual_bases),\n                )\n            )::t_old.virtual_bases::1',
-        'abicheck/diff_types.py::_diff_type_bases::virtual_bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_VIRTUAL_CHANGED,\n                symbol=name,\n                name=name,\n                detail="; ".join(desc_parts),\n                old_value=str(sorted(t_old.virtual_bases)),\n                new_value=str(sorted(t_new.virtual_bases)),\n            )\n        )::t_new.virtual_bases::1',
-        'abicheck/diff_types.py::_diff_type_bases::virtual_bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_VIRTUAL_CHANGED,\n                symbol=name,\n                name=name,\n                detail="; ".join(desc_parts),\n                old_value=str(sorted(t_old.virtual_bases)),\n                new_value=str(sorted(t_new.virtual_bases)),\n            )\n        )::t_old.virtual_bases::1',
-        "abicheck/diff_types.py::_diff_type_bases::virtual_bases::set(t_new.virtual_bases)::t_new.virtual_bases::1",
-        "abicheck/diff_types.py::_diff_type_bases::virtual_bases::set(t_old.virtual_bases)::t_old.virtual_bases::1",
-        'abicheck/diff_types.py::_diff_type_vtable::vtable::f"vtable reordered: {name}"\n        if Counter(t_old.vtable) == Counter(t_new.vtable)\n        else f"vtable changed: {name}"::t_new.vtable::1',
-        'abicheck/diff_types.py::_diff_type_vtable::vtable::f"vtable reordered: {name}"\n        if Counter(t_old.vtable) == Counter(t_new.vtable)\n        else f"vtable changed: {name}"::t_old.vtable::1',
-        "abicheck/diff_types.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_new.vtable::1",
-        "abicheck/diff_types.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_new.vtable::2",
-        "abicheck/diff_types.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_old.vtable::1",
-        "abicheck/diff_types.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_old.vtable::2",
-        'abicheck/diff_types.py::_diff_type_vtable::vtable::make_change(\n        ChangeKind.TYPE_VTABLE_CHANGED,\n        symbol=name,\n        description=description,\n        old_value=", ".join(t_old.vtable),\n        new_value=", ".join(t_new.vtable),\n    )::t_new.vtable::1',
-        'abicheck/diff_types.py::_diff_type_vtable::vtable::make_change(\n        ChangeKind.TYPE_VTABLE_CHANGED,\n        symbol=name,\n        description=description,\n        old_value=", ".join(t_old.vtable),\n        new_value=", ".join(t_new.vtable),\n    )::t_old.vtable::1',
-        "abicheck/diff_types.py::_diff_type_vtable::vtable::t_old.vtable == t_new.vtable::t_new.vtable::1",
-        "abicheck/diff_types.py::_diff_type_vtable::vtable::t_old.vtable == t_new.vtable::t_old.vtable::1",
         "abicheck/diff_types.py::_new_field_change_kind::virtual_bases::bool(t_new.vtable or t_new.virtual_bases)::t_new.virtual_bases::1",
         "abicheck/diff_types.py::_new_field_change_kind::vtable::bool(t_new.vtable or t_new.virtual_bases)::t_new.vtable::1",
-        "abicheck/diff_types.py::_vtable_transition_is_evidenced::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_new.virtual_bases::1",
-        "abicheck/diff_types.py::_vtable_transition_is_evidenced::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_old.virtual_bases::1",
-        "abicheck/diff_types.py::_vtable_transition_is_evidenced::vtable::t_old.vtable and t_new.vtable::t_new.vtable::1",
-        "abicheck/diff_types.py::_vtable_transition_is_evidenced::vtable::t_old.vtable and t_new.vtable::t_old.vtable::1",
-        "abicheck/diff_types.py::_vtable_transition_rests_on_unresolved_evidence::bases::list(t_old.bases) != list(t_new.bases)::t_new.bases::1",
-        "abicheck/diff_types.py::_vtable_transition_rests_on_unresolved_evidence::bases::list(t_old.bases) != list(t_new.bases)::t_old.bases::1",
-        "abicheck/diff_types.py::_vtable_transition_rests_on_unresolved_evidence::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_new.virtual_bases::1",
-        "abicheck/diff_types.py::_vtable_transition_rests_on_unresolved_evidence::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_old.virtual_bases::1",
-        "abicheck/diff_types.py::_vtable_transition_rests_on_unresolved_evidence::vtable::t_old.vtable and t_new.vtable::t_new.vtable::1",
-        "abicheck/diff_types.py::_vtable_transition_rests_on_unresolved_evidence::vtable::t_old.vtable and t_new.vtable::t_old.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_POSITION_CHANGED,\n                symbol=name,\n                name=name,\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n                entity_id=t_old.entity_id or t_new.entity_id,\n            )\n        )::t_new.bases::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_POSITION_CHANGED,\n                symbol=name,\n                name=name,\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n                entity_id=t_old.entity_id or t_new.entity_id,\n            )\n        )::t_old.bases::1",
+        'abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.TYPE_BASE_CHANGED,\n                symbol=name,\n                description=f"Base classes changed: {name}",\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n                entity_id=t_old.entity_id or t_new.entity_id,\n            )\n        )::t_new.bases::1',
+        'abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::changes.append(\n            make_change(\n                ChangeKind.TYPE_BASE_CHANGED,\n                symbol=name,\n                description=f"Base classes changed: {name}",\n                old_value=str(t_old.bases),\n                new_value=str(t_new.bases),\n                entity_id=t_old.entity_id or t_new.entity_id,\n            )\n        )::t_old.bases::1',
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::old_bases_set == new_bases_set and t_old.bases != t_new.bases::t_new.bases::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::old_bases_set == new_bases_set and t_old.bases != t_new.bases::t_old.bases::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::set(t_new.bases)::t_new.bases::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::bases::set(t_old.bases)::t_old.bases::1",
+        'abicheck/diff_types_bases_vtable.py::_diff_type_bases::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.TYPE_BASE_CHANGED,\n                    symbol=name,\n                    description=f"Virtual base classes changed: {name}",\n                    old_value=str(t_old.virtual_bases),\n                    new_value=str(t_new.virtual_bases),\n                    entity_id=t_old.entity_id or t_new.entity_id,\n                )\n            )::t_new.virtual_bases::1',
+        'abicheck/diff_types_bases_vtable.py::_diff_type_bases::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.TYPE_BASE_CHANGED,\n                    symbol=name,\n                    description=f"Virtual base classes changed: {name}",\n                    old_value=str(t_old.virtual_bases),\n                    new_value=str(t_new.virtual_bases),\n                    entity_id=t_old.entity_id or t_new.entity_id,\n                )\n            )::t_old.virtual_bases::1',
+        'abicheck/diff_types_bases_vtable.py::_diff_type_bases::virtual_bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_VIRTUAL_CHANGED,\n                symbol=name,\n                name=name,\n                detail="; ".join(desc_parts),\n                old_value=str(sorted(t_old.virtual_bases)),\n                new_value=str(sorted(t_new.virtual_bases)),\n                entity_id=t_old.entity_id or t_new.entity_id,\n            )\n        )::t_new.virtual_bases::1',
+        'abicheck/diff_types_bases_vtable.py::_diff_type_bases::virtual_bases::changes.append(\n            make_change(\n                ChangeKind.BASE_CLASS_VIRTUAL_CHANGED,\n                symbol=name,\n                name=name,\n                detail="; ".join(desc_parts),\n                old_value=str(sorted(t_old.virtual_bases)),\n                new_value=str(sorted(t_new.virtual_bases)),\n                entity_id=t_old.entity_id or t_new.entity_id,\n            )\n        )::t_old.virtual_bases::1',
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::virtual_bases::set(t_new.virtual_bases)::t_new.virtual_bases::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_bases::virtual_bases::set(t_old.virtual_bases)::t_old.virtual_bases::1",
+        'abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::f"vtable reordered: {name}"\n        if Counter(t_old.vtable) == Counter(t_new.vtable)\n        else f"vtable changed: {name}"::t_new.vtable::1',
+        'abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::f"vtable reordered: {name}"\n        if Counter(t_old.vtable) == Counter(t_new.vtable)\n        else f"vtable changed: {name}"::t_old.vtable::1',
+        "abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_new.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_new.vtable::2",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_old.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::len(t_old.vtable) == len(t_new.vtable) and all(\n        vtable_slot_is_override_reuse(\n            old_entry, new_entry, old_funcs, new_funcs, old_types, new_types\n        )\n        for old_entry, new_entry in zip(t_old.vtable, t_new.vtable)\n    )::t_old.vtable::2",
+        'abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::make_change(\n        ChangeKind.TYPE_VTABLE_CHANGED,\n        symbol=name,\n        description=description,\n        old_value=", ".join(t_old.vtable),\n        new_value=", ".join(t_new.vtable),\n        entity_id=t_old.entity_id or t_new.entity_id,\n    )::t_new.vtable::1',
+        'abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::make_change(\n        ChangeKind.TYPE_VTABLE_CHANGED,\n        symbol=name,\n        description=description,\n        old_value=", ".join(t_old.vtable),\n        new_value=", ".join(t_new.vtable),\n        entity_id=t_old.entity_id or t_new.entity_id,\n    )::t_old.vtable::1',
+        "abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::t_old.vtable == t_new.vtable::t_new.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_diff_type_vtable::vtable::t_old.vtable == t_new.vtable::t_old.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_is_evidenced::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_new.virtual_bases::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_is_evidenced::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_old.virtual_bases::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_is_evidenced::vtable::t_old.vtable and t_new.vtable::t_new.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_is_evidenced::vtable::t_old.vtable and t_new.vtable::t_old.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_rests_on_unresolved_evidence::bases::list(t_old.bases) != list(t_new.bases)::t_new.bases::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_rests_on_unresolved_evidence::bases::list(t_old.bases) != list(t_new.bases)::t_old.bases::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_rests_on_unresolved_evidence::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_new.virtual_bases::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_rests_on_unresolved_evidence::virtual_bases::list(t_old.virtual_bases) != list(t_new.virtual_bases)::t_old.virtual_bases::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_rests_on_unresolved_evidence::vtable::t_old.vtable and t_new.vtable::t_new.vtable::1",
+        "abicheck/diff_types_bases_vtable.py::_vtable_transition_rests_on_unresolved_evidence::vtable::t_old.vtable and t_new.vtable::t_old.vtable::1",
         "abicheck/diff_vtable_layout.py::_diff_vtable_layout::bases::o.bases == n.bases and o.virtual_bases == n.virtual_bases::n.bases::1",
         "abicheck/diff_vtable_layout.py::_diff_vtable_layout::bases::o.bases == n.bases and o.virtual_bases == n.virtual_bases::o.bases::1",
-        'abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.VIRTUAL_BASE_OFFSET_CHANGED,\n                    symbol=name,\n                    name=name,\n                    old=", ".join(o.virtual_bases),\n                    new=", ".join(n.virtual_bases),\n                )\n            )::n.virtual_bases::1',
-        'abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.VIRTUAL_BASE_OFFSET_CHANGED,\n                    symbol=name,\n                    name=name,\n                    old=", ".join(o.virtual_bases),\n                    new=", ".join(n.virtual_bases),\n                )\n            )::o.virtual_bases::1',
+        'abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.VIRTUAL_BASE_OFFSET_CHANGED,\n                    symbol=name,\n                    name=name,\n                    old=", ".join(o.virtual_bases),\n                    new=", ".join(n.virtual_bases),\n                    entity_id=o.entity_id or n.entity_id,\n                )\n            )::n.virtual_bases::1',
+        'abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::changes.append(\n                make_change(\n                    ChangeKind.VIRTUAL_BASE_OFFSET_CHANGED,\n                    symbol=name,\n                    name=name,\n                    old=", ".join(o.virtual_bases),\n                    new=", ".join(n.virtual_bases),\n                    entity_id=o.entity_id or n.entity_id,\n                )\n            )::o.virtual_bases::1',
         "abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::len(o.virtual_bases) > 1\n            and set(o.virtual_bases) == set(n.virtual_bases)\n            and o.virtual_bases != n.virtual_bases::n.virtual_bases::1",
         "abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::len(o.virtual_bases) > 1\n            and set(o.virtual_bases) == set(n.virtual_bases)\n            and o.virtual_bases != n.virtual_bases::n.virtual_bases::2",
         "abicheck/diff_vtable_layout.py::_diff_vtable_layout::virtual_bases::len(o.virtual_bases) > 1\n            and set(o.virtual_bases) == set(n.virtual_bases)\n            and o.virtual_bases != n.virtual_bases::o.virtual_bases::1",
