+++ tests/test_cli_coverage_extra.py
@@ -95,8 +95,13 @@ class TestDumpNativeBinary:
         # global -- so that is the only place patching it has ever had an
         # effect. The previous `abicheck.cli.` target was already inert; the
         # facade split surfaced it by making the attribute stop existing.
+        # ADR-063 Phase 1: the real PE dump now executes through
+        # `execute_dump_request` (`abicheck.service_dump_native._dump_pe`),
+        # not `commands.dump._dump_native_binary` -- see
+        # `test_compile_context_parity.py`'s `_capture_dump_pe` for the same
+        # patch target, established by that migration.
         with patch("abicheck.cli_resolve._detect_binary_format", return_value="pe"), \
-             patch("abicheck.frontends.cli.commands.dump._dump_native_binary", return_value=mock_snap):
+             patch("abicheck.service_dump_native._dump_pe", return_value=mock_snap):
             runner = CliRunner()
             result = runner.invoke(main, [
                 "dump", str(pe_file), "--version", "1.0", "--follow-deps",
@@ -118,9 +123,10 @@ class TestDumpNativeBinary:
         only the ELF dump path forwarded the opt-in flag and PE/Mach-O
         silently ignored ``--header-graph`` (Codex review); that class of bug
         can no longer occur since there is nothing left to forward
-        selectively. This also confirms ``_dump_native_binary`` is no longer
-        called with ``header_graph``/``header_graph_includes`` kwargs at
-        all."""
+        selectively. This also confirms ``_dump_pe`` (ADR-063 Phase 1: the
+        real PE dump target, since the migration onto
+        ``execute_dump_request``) is not called with
+        ``header_graph``/``header_graph_includes`` kwargs at all."""
         from abicheck.model import AbiSnapshot
 
         pe_file = tmp_path / "test.dll"
@@ -134,7 +140,7 @@ class TestDumpNativeBinary:
             return mock_snap
 
         with patch("abicheck.cli_resolve._detect_binary_format", return_value="pe"), \
-             patch("abicheck.frontends.cli.commands.dump._dump_native_binary", side_effect=_fake_dump_native):
+             patch("abicheck.service_dump_native._dump_pe", side_effect=_fake_dump_native):
             runner = CliRunner()
             result = runner.invoke(main, [
                 "dump", str(pe_file), "--version", "1.0",
@@ -156,7 +162,7 @@ class TestDumpNativeBinary:
         mock_snap = AbiSnapshot(library="test.dll", version="1.0", platform="pe")
 
         with patch("abicheck.cli_resolve._detect_binary_format", return_value="pe"), \
-             patch("abicheck.frontends.cli.commands.dump._dump_native_binary", return_value=mock_snap):
+             patch("abicheck.service_dump_native._dump_pe", return_value=mock_snap):
             runner = CliRunner()
             result = runner.invoke(main, [
                 "dump", str(pe_file), "--version", "1.0",
@@ -185,7 +191,7 @@ class TestDumpNativeBinary:
         )
 
         with patch("abicheck.cli_resolve._detect_binary_format", return_value="pe"), \
-             patch("abicheck.frontends.cli.commands.dump._dump_native_binary", return_value=mock_snap):
+             patch("abicheck.service_dump_native._dump_pe", return_value=mock_snap):
             runner = CliRunner()
             result = runner.invoke(main, [
                 "dump", str(pe_file), "--version", "1.0",
